NoImageProvider and NoImageResolver as one service
- a logic from an interface `INoImageProvider` has been moved to an interface `INoImageResolver`
- a logic from a class `NoImageProvider` has been moved to a class `NoImageResolver`
- removed the interface `INoImageProvider`
- removed the class `NoImageProvider`
- renamed an extension option `no_image_rules` to `no_image_patterns`
