+++ src/NoImage/INoImageResolver.php
@@ -9,6 +9,20 @@ use SixtyEightPublishers;
 interface INoImageResolver
 {
 	/**
+	 * @param string|NULL $name
+	 *
+	 * @return \SixtyEightPublishers\ImageStorage\ImageInfo
+	 */
+	public function getNoImage(?string $name = NULL): SixtyEightPublishers\ImageStorage\ImageInfo;
+
+	/**
+	 * @param string $path
+	 *
+	 * @return bool
+	 */
+	public function isNoImage(string $path): bool;
+
+	/**
 	 * @param string $path
 	 *
 	 * @return \SixtyEightPublishers\ImageStorage\ImageInfo
